Change the timekeeping method for logging CF variables.
Why?

Ans). With the goal to log mocap data in addition to the CF variables, I noticed that
the two (CF logging subsystem) and Qualisys use different 'clocks' -- The Crazyflie
sends log data back with a timestamp based on the microcontroller clock. Plotting these two
kinds of data on the same time axis is difficult because of this.

Solution:

Replace the CF timestamps with a call to 'time.time()'.